Rewrite the control structure.
There are now skip forwards and skip backwards buttons, which seek a default 15 seconds
forwards or back from the current position (this can be customised with the skipTime
option).

Volume up and down buttons have also been added, so that users without a mouse or
pointing device can change the volume settings. These increase or decrease the volume
by 25%.